added confirm to del button

+++ public/javascript/deletePost.js
@@ -1,14 +1,19 @@
 async function reply_click(id){
   
-    const response = await fetch(`/api/posts/${id}`, {
-        method: 'DELETE'
-    });
+    const safeguard = await confirm("Are you sure that you would like to delete this post?");
+    if (safeguard == true) {
+        const response = await fetch(`/api/posts/${id}`, {
+            method: 'DELETE'
+        });
+    
+        if(response.ok) {
+        document.location.replace('/dashboard');
+        }
+         else {
+        alert(response.statusText);
+        }
+    }
 
-    if(response.ok) {
-    document.location.replace('/dashboard');
-    }
-     else {
-    alert(response.statusText);
-    }
+    
     
 }
